refactoring - moved http.client related things into isolated package
