+++ pkg/httpclient/tls.go
@@ -1,0 +1,35 @@
+package httpclient
+
+import (
+	"crypto/tls"
+	"crypto/x509"
+	"errors"
+
+	"github.com/grafana/grafana-infinity-datasource/pkg/models"
+)
+
+func GetTLSConfigFromSettings(settings models.InfinitySettings) (*tls.Config, error) {
+	tlsConfig := &tls.Config{
+		InsecureSkipVerify: settings.InsecureSkipVerify,
+		ServerName:         settings.ServerName,
+	}
+	if settings.TLSClientAuth {
+		if settings.TLSClientCert == "" || settings.TLSClientKey == "" {
+			return nil, errors.New("invalid Client cert or key")
+		}
+		cert, err := tls.X509KeyPair([]byte(settings.TLSClientCert), []byte(settings.TLSClientKey))
+		if err != nil {
+			return nil, err
+		}
+		tlsConfig.Certificates = []tls.Certificate{cert}
+	}
+	if settings.TLSAuthWithCACert && settings.TLSCACert != "" {
+		caPool := x509.NewCertPool()
+		ok := caPool.AppendCertsFromPEM([]byte(settings.TLSCACert))
+		if !ok {
+			return nil, errors.New("invalid TLS CA certificate")
+		}
+		tlsConfig.RootCAs = caPool
+	}
+	return tlsConfig, nil
+}
